move last openshift-controller-manager to staging
