+++ hack/verify-imports.sh
@@ -78,11 +78,15 @@ RC=0
 print_forbidden_imports oauth-server k8s.io/kubernetes/pkg/apis || RC=1
 print_forbidden_imports oc k8s.io/kubernetes/pkg || RC=1
 print_forbidden_imports openshift-apiserver k8s.io/kubernetes/pkg/apis || RC=1
-print_forbidden_imports openshift-controller-manager k8s.io/kubernetes/pkg/api/legacyscheme \
+print_forbidden_imports openshift-controller-manager k8s.io/kubernetes/cmd/controller-manager/app \
+  k8s.io/kubernetes/pkg/api/legacyscheme \
   k8s.io/kubernetes/pkg/api/testing \
   k8s.io/kubernetes/pkg/apis \
+  k8s.io/kubernetes/pkg/client/metrics/prometheus \
   k8s.io/kubernetes/pkg/controller \
   k8s.io/kubernetes/pkg/credentialprovider \
+  k8s.io/kubernetes/pkg/kubectl/cmd/util \
+  k8s.io/kubernetes/pkg/kubectl/util/templates \
   k8s.io/kubernetes/pkg/quota/v1 \
   k8s.io/kubernetes/pkg/registry/core/secret \
   k8s.io/kubernetes/pkg/registry/core/service || RC=1
