+++ staging/src/github.com/openshift/openshift-controller-manager/pkg/cmd/controller/apiserver_authenticator.go
