+++ staging/src/github.com/openshift/openshift-controller-manager/pkg/cmd/controller/image.go
