+++ staging/src/github.com/openshift/openshift-controller-manager/pkg/cmd/controller/template.go
