+++ staging/src/github.com/openshift/openshift-controller-manager/pkg/cmd/openshift-controller-manager/cmd.go
